# swagger-spec.json
{"openapi":"3.0.0","paths":{"/payment/success":{"post":{"operationId":"PaymentController_paymentSuccess","parameters":[],"responses":{"201":{"description":""}},"tags":["Payment"]}},"/calendar":{"post":{"operationId":"CalendarController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCalendarDto"}}}},"responses":{"201":{"description":""}},"tags":["Calendar"]},"get":{"operationId":"CalendarController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Calendar"]}},"/calendar/freebusy":{"get":{"operationId":"CalendarController_getFreeBusy","parameters":[],"responses":{"200":{"description":""}},"tags":["Calendar"]}},"/calendar/{id}":{"get":{"operationId":"CalendarController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Calendar"]},"patch":{"operationId":"CalendarController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCalendarDto"}}}},"responses":{"200":{"description":""}},"tags":["Calendar"]},"delete":{"operationId":"CalendarController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Calendar"]}},"/availability":{"get":{"operationId":"AvailabilityController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Availability"]}},"/availability/workslots":{"get":{"operationId":"AvailabilityController_findWorkSlots","parameters":[],"responses":{"200":{"description":""}},"tags":["Availability"]}},"/availability/current":{"get":{"operationId":"AvailabilityController_findCurrent","parameters":[],"responses":{"200":{"description":""}},"tags":["Availability"]}},"/availability/{id}":{"get":{"operationId":"AvailabilityController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Availability"]}},"/free-slot":{"get":{"operationId":"FreeSlotController_getFreeSlots","parameters":[{"name":"guestTimezone","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Free Time Slots"]}}},"info":{"title":"AprendeCoding API","description":"Welcome to the AprendeCoding API reference.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateCalendarDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name","description"]},"UpdateCalendarDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}}}}}